requirements.txt
APScheduler==3.3.1
click==6.7
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.5
django-common-helpers==0.9.0
django-cron==0.5.0
funcsigs==1.0.2
future==0.16.0
futures==3.0.5
gunicorn==19.6.0
psycopg2==2.6.2
pytz==2016.10
PyYAML==3.12
redis==2.10.5
requests==2.12.5
rq==0.7.1
six==1.10.0
static3==0.7.0
tzlocal==1.3
uber-rides==0.3.1
urllib3==1.20
whitenoise==3.2.3
